% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{abort_no_sample_variable}
\alias{abort_no_sample_variable}
\title{Throw an error for a variable that does not exist in the sample data of a
\code{SummarizedExperiment}}
\usage{
abort_no_sample_variable(var, se)
}
\arguments{
\item{var}{The sample data variable that does not exist in the
\code{SummarizedExperiment}}

\item{se}{The \code{SummarizedExperiment} that does not contain the sample data
variable}
}
\value{
An \code{rlang_error} of custom class \code{error_no_sample_variable}
}
\description{
Throw an error for a variable that does not exist in the sample data of a
\code{SummarizedExperiment}
}
