% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-cor_dist.R
\name{compute_cor}
\alias{compute_cor}
\title{Compute pairwise sample--sample correlations for a \code{SummarizedExperiment}
\code{assay}}
\usage{
compute_cor(se, assay, method)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{assay}{A character vector indicating the assay to compute correlation
for.}

\item{method}{A character vector indicating the correlation method. Passed to
\link[stats:cor]{stats::cor}. Also sets a name for the heatmap unless a \code{name} argument is
supplied.}
}
\value{
A matrix of pairwise sample correlations.
}
\description{
Compute pairwise sample--sample correlations for a \code{SummarizedExperiment}
\code{assay}
}
