% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-cor_dist.R
\name{compute_dist}
\alias{compute_dist}
\title{Compute pairwise sample--sample distances for a \code{SummarizedExperiment}
\code{assay}}
\usage{
compute_dist(
  se,
  assay,
  method = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski")
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{assay}{A character vector indicating the assay to compute correlation
for.}

\item{method}{A character vector passed to \link[stats:dist]{stats::dist}. Also sets a name for the heatmap unless a \code{name} argument is
supplied.}
}
\value{
A matrix of pairwise sample--sample distances.
}
\description{
Compute pairwise sample--sample distances for a \code{SummarizedExperiment}
\code{assay}
}
