% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_se.R
\name{construct_se}
\alias{construct_se}
\title{Construct a \code{SummarizedExperiment} from a counts matrix and sample data data
frame}
\usage{
construct_se(counts, samples, library_size_var = "library_size", ...)
}
\arguments{
\item{counts}{An integer matrix with genes as rows and samples as columns.
Row names are gene names. Column names are sample names.}

\item{samples}{A data frame of samples data. Row names are sample names.}

\item{library_size_var}{A character vector name for the library size
variable. If \code{NULL}, library size is not computed.}

\item{...}{Additional arguments passed to
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}}.}
}
\value{
A \code{SummarizedExperiment} object. \code{counts} assay contains counts data.
}
\description{
\code{construct_se()} is essentially a wrapper around
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} that offers additional
validation of sample order in the counts matrix and samples data as
required to construct the \code{SummarizedExperiment} object.

\code{construct_se()} subsets the counts matrix and column data by the sorted
intersection of counts matrix column names and sample data row names, then
constructs a \code{SummarizedExperiment} object from the subset of the counts
matrix and sample data.

\code{construct_se()} by default adds a variable to the sample data for library
size, that is, the column sums of the counts matrix. Controlled by
the \code{library_size_var} argument.
}
