% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_genes.R
\name{count_experiment_detected_genes}
\alias{count_experiment_detected_genes}
\title{Count detected genes for a whole experiment}
\usage{
count_experiment_detected_genes(dds, threshold = 5, covariates = NULL)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{threshold}{Threshold for minimum normalized counts.}

\item{covariates}{A vector of covariates found in \code{dds} to count the number of samples within.
If \code{NULL}, the covariates from the \code{dds} design formula will be used.}
}
\value{
Count of experiment-wide detected genes.
}
\description{
Count the number of genes with normalized counts above a specified threshold in the smallest condition set.
}
