% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_genes.R
\name{count_library_detected_genes}
\alias{count_library_detected_genes}
\title{Count detected genes for each library}
\usage{
count_library_detected_genes(dds, threshold = 5)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{threshold}{Threshold for minimum normalized counts.}
}
\value{
A named vector of counts of detected genes for each sample.
}
\description{
Count detected genes for each library
}
