% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_genes.R
\name{count_samples_each_condition}
\alias{count_samples_each_condition}
\title{Count the number of samples in each condition}
\usage{
count_samples_each_condition(dds, covariates = NULL)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{covariates}{A vector of covariates found in \code{dds} to count the number of samples within.
If \code{NULL}, the covariates from the \code{dds} design formula will be used.}
}
\value{
A data frame with a count of the number of samples in each condition.
}
\description{
Count the number of samples in each condition
}
