% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_exclusions.R
\name{drop_excluded_samples}
\alias{drop_excluded_samples}
\title{Drop excluded samples from a \code{SummarizedExperiment}}
\usage{
drop_excluded_samples(se, sample_exclusions)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object}

\item{sample_exclusions}{A \code{list} of key:value pairs mapping reason for
exclusion to samples excluded.}
}
\value{
A \code{SummarizedExperiment} object without the columns that overlap
sample names found in \code{sample_exclusions}
}
\description{
Drop excluded samples from a \code{SummarizedExperiment}
}
