% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary.R
\name{extract_results_summary_data}
\alias{extract_results_summary_data}
\title{Extract summary data for \code{DESeqResults}}
\usage{
extract_results_summary_data(object)
}
\arguments{
\item{object}{A \code{DESeqResults} object.}
}
\value{
A \code{DataFrame} that closely mimics the output of \code{DESeq2::summary()}.
}
\description{
This is essentially a rewrite of \code{DESeq2::summary()} that returns
a useful \code{S4Vectors::DF} object instead of just printing summary info.
}
