% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary.R
\name{extract_shrunken_results}
\alias{extract_shrunken_results}
\title{Wrap together \code{\link[DESeq2:results]{DESeq2::results()}} and \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} for convenience}
\usage{
extract_shrunken_results(
  dds,
  name,
  alpha = 0.05,
  lfcThreshold = 0,
  type = "apeglm",
  quiet = TRUE
)
}
\arguments{
\item{dds}{A \code{DESeqDataSet} object}

\item{name}{The name of the coefficient/comparison to extract results for
(see \code{\link[DESeq2:results]{DESeq2::resultsNames()}}). This is passed to the \code{name} argument of
\code{\link[DESeq2:results]{DESeq2::results()}} and the \code{coef} argument of \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.}

\item{alpha}{the significance cutoff used for optimizing the independent
filtering (by default 0.1). If the adjusted p-value cutoff (FDR) will be a
value other than 0.1, \code{alpha} should be set to that value.}

\item{lfcThreshold}{a non-negative value which specifies a log2 fold change
threshold. The default value is 0, corresponding to a test that
the log2 fold changes are equal to zero. The user can
specify the alternative hypothesis using the \code{altHypothesis} argument,
which defaults to testing
for log2 fold changes greater in absolute value than a given threshold.
If \code{lfcThreshold} is specified,
the results are for Wald tests, and LRT p-values will be overwritten.}

\item{type}{\code{"apeglm"} is the adaptive Student's t prior shrinkage estimator from the 'apeglm' package;
\code{"ashr"} is the adaptive shrinkage estimator from the 'ashr' package,
using a fitted mixture of normals prior
- see the Stephens (2016) reference below for citation;
\code{"normal"} is the 2014 DESeq2 shrinkage estimator using a Normal prior;}

\item{quiet}{whether to print messages}
}
\value{
A \code{DESeqResults} object with shrunken log2 fold changes.
}
\description{
A common workflow in \code{DESeq2} involves extracting a \code{DESeqResults}
object with \code{\link[DESeq2:results]{DESeq2::results()}} followed by shrinkage of log2 fold changes
with \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}. These are kept as separate functions within
\code{DESeq2} so that additional parameters can be passed to the functions for
shrinkage estimation. However, I do not often do this, so I wrap these two
functions together for convenience here.
}
