% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{get_alpha}
\alias{get_alpha}
\title{Return alpha value of a \code{TestSpecs} object}
\usage{
get_alpha(test_specs)
}
\arguments{
\item{test_specs}{A \code{TestSpecs} object}
}
\value{
A numeric alpha threshold
}
\description{
Return alpha value of a \code{TestSpecs} object
}
