% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Get all contrasts from a \code{TestSpecs} object}
\usage{
get_contrasts(test_specs)
}
\arguments{
\item{test_specs}{A \code{TestSpecs} object}
}
\value{
A named character vector. Names are contrast names, values are contrasts.
}
\description{
Get all contrasts from a \code{TestSpecs} object
}
