% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{get_contrasts_table}
\alias{get_contrasts_table}
\title{Get all contrasts from a \code{TestSpecs} object as a single data.frame}
\usage{
get_contrasts_table(test_specs)
}
\arguments{
\item{test_specs}{A \code{TestSpecs} object}
}
\value{
A data.frame with contrast names, contrasts, and descriptions.
}
\description{
Get all contrasts from a \code{TestSpecs} object as a single data.frame
}
