% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_genes.R
\name{get_design_covariates}
\alias{get_design_covariates}
\title{Get covariates from design formulas}
\usage{
get_design_covariates(dds)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}
}
\value{
A character vector of covariates found in the design formula.
}
\description{
Get covariates from design formulas
}
