% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{get_formula}
\alias{get_formula}
\title{Get a formula from the metadata of a \code{SummarizedExperiment} by its name}
\usage{
get_formula(se, name)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{name}{A character name for the formula.}
}
\value{
The requested formula
}
\description{
Get a formula from the metadata of a \code{SummarizedExperiment} by its name
}
