% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{get_lfc_threshold}
\alias{get_lfc_threshold}
\title{Return log2 fold change threshold of a \code{TestSpecs} object}
\usage{
get_lfc_threshold(test_specs)
}
\arguments{
\item{test_specs}{A \code{TestSpecs} object}
}
\value{
A numeric log2 fold change threshold
}
\description{
Return log2 fold change threshold of a \code{TestSpecs} object
}
