% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{get_modeling_all_variables}
\alias{get_modeling_all_variables}
\title{Get all modeling variables from a \code{SummarizedExperiment}}
\usage{
get_modeling_all_variables(se)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}
}
\value{
A character vector of all modeling variables.
}
\description{
Get all modeling variables from a \code{SummarizedExperiment}
}
