% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{get_modeling_variable}
\alias{get_modeling_variable}
\title{Get a modeling variable in the metadata of a \code{SummarizedExperiment}}
\usage{
get_modeling_variable(se, variable = c("interest", "nuisance", "screening"))
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{variable}{A character vector for which variable to set}
}
\value{
A character vector of modeling variables.
}
\description{
Get a modeling variable in the metadata of a \code{SummarizedExperiment}
}
