% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{get_prcomp}
\alias{get_prcomp}
\title{Access prcomp from a PCA DFrame}
\usage{
get_prcomp(pca_dframe)
}
\arguments{
\item{pca_dframe}{A \code{Dframe} of PCA data}
}
\value{
A \code{prcomp} object
}
\description{
An accessor method for getting the \code{prcomp} object from a PCA DFrame
such as that produced by \code{\link[=perform_pca]{perform_pca()}}
}
