% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{get_samples_data}
\alias{get_samples_data}
\title{Get samples data (column data) from a \code{SummarizedExperiment}}
\usage{
get_samples_data(se, rownames_var = "sample")
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{rownames_var}{A character vector for the name of the variable to store
row names as. If \code{NULL}, row names stay in the row names and no new
variable in the data table is created.}
}
\value{
A data.frame of samples data.
}
\description{
Get samples data (column data) from a \code{SummarizedExperiment}
}
