% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{get_tests_table}
\alias{get_tests_table}
\title{Get all tests from a \code{TestSpecs} object as a single data.frame}
\usage{
get_tests_table(test_specs)
}
\arguments{
\item{test_specs}{A \code{TestSpecs} object}
}
\value{
A data.frame with test names, contrast names, and descriptions.
}
\description{
For any tests with more than one contrast name, contrast names are returned
as a single pipe ('|') delimited string.
}
