% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{get_variance_explained}
\alias{get_variance_explained}
\title{Access variance explained from a PCA DFrame}
\usage{
get_variance_explained(pca_dframe)
}
\arguments{
\item{pca_dframe}{A \code{Dframe} of PCA data}
}
\value{
A data frame with the proportion of variance explained and cumulative
proportion of variance explained for each principal component.
}
\description{
An accessor method for getting the variance explained data from
a PCA DFrame such as that produced by \code{\link[=perform_pca]{perform_pca()}}
}
