% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-utils.R
\name{mkdir_p}
\alias{mkdir_p}
\title{Create directory to path}
\usage{
mkdir_p(path)
}
\arguments{
\item{path}{A character path to a directory}
}
\value{
A logical indicating if the directory creation succeeded
}
\description{
The goal is to mimic the behavior of \code{mkdir -p} command. That is, create the
full directory path recursively without any errors or warnings if the
directory already exists.
}
