% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{new_TestSpecs}
\alias{new_TestSpecs}
\title{Low level \code{TestSpecs} constructor}
\usage{
new_TestSpecs(tests, contrasts, alpha, lfc_threshold)
}
\arguments{
\item{tests}{A list of tests data}

\item{contrasts}{A list of contrasts data}

\item{alpha}{A numeric alpha threshold for differential features}

\item{lfc_threshold}{A numeric log2 fold change threshold for differential features}
}
\value{
A \code{TestSpecs} object
}
\description{
Low level \code{TestSpecs} constructor
}
