% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform PCA and return data for transformed data}
\usage{
perform_pca(object, ntop = 500)
}
\arguments{
\item{object}{a \code{\link[DESeq2]{DESeqTransform}} object, with data in \code{assay(x)},
produced for example by either \code{\link[DESeq2]{rlog}} or
\code{\link[DESeq2]{varianceStabilizingTransformation}}.}

\item{ntop}{number of top genes to use for principal components,
selected by highest row variance}
}
\value{
A \code{DFrame} with a \code{prcomp} object in the metadata.
}
\description{
Perform PCA and return data for transformed data
}
