% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-variance_partition.R
\name{perform_variance_partition_analysis}
\alias{perform_variance_partition_analysis}
\title{Performance variance partitioning analysis}
\usage{
perform_variance_partition_analysis(se, assay, formula_name, ...)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{assay}{A character vector for the assay from \code{SummarizedExperiment} to
use as input to the PCA.}

\item{formula_name}{A character name for the formula in formula models slot
of \code{SummarizedExperiment} metadata.}

\item{...}{Additional parameters passed to
\link[variancePartition:fitExtractVarPartModel-method]{variancePartition::fitExtractVarPartModel}. The \code{exprObj}, \code{formula}, and
\code{data} arguments are already set from the \code{SummarizedExperiment}.}
}
\value{
A \code{varPartResults} from \link[variancePartition:fitExtractVarPartModel-method]{variancePartition::fitExtractVarPartModel}.
}
\description{
Performance variance partitioning analysis
}
