% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_biplot}
\alias{plot_biplot}
\title{Make a biplot for a PCA DFrame}
\usage{
plot_biplot(pca_df, x = PC1, y = PC2, point_mapping = NULL)
}
\arguments{
\item{pca_df}{A \code{Dframe} of PCA data}

\item{x}{Principal component to plot along x axis}

\item{y}{Prinicpal component to plot along y axis}

\item{point_mapping}{A mapping aesthetic to pass to \link[ggplot2:geom_point]{ggplot2::geom_point}.
Created by \link[ggplot2:aes]{ggplot2::aes}.}
}
\value{
A ggplot object biplot.
}
\description{
Make a biplot for a PCA DFrame
}
