% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-sample_data.R
\name{plot_canonical_correlation}
\alias{plot_canonical_correlation}
\title{Plot canonical correlation heatmap for sample data of specific variable
classes from \code{SummarizedExperiment}}
\usage{
plot_canonical_correlation(
  se,
  modeling_variable_class = c("all", "explanatory")
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{modeling_variable_class}{A character for the class of modeling
variables to plot canonical correlation for.}
}
\value{
A \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap} of canonical correlation values for the
sample data variables for a specified class of modeling variables.
}
\description{
Plot canonical correlation heatmap for sample data of specific variable
classes from \code{SummarizedExperiment}
}
