% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-cor_dist.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{Compute and plot pairwise sample--sample correlations for a
\code{SummarizedExperiment} \code{assay}}
\usage{
plot_cor_heatmap(
  se,
  assay,
  method = c("pearson", "kendall", "spearman"),
  ...,
  annotations = c("both", "top", "left", "none"),
  diagonal = c("remove", "as_is")
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{assay}{A character vector indicating the assay to compute correlation
for.}

\item{method}{A character vector indicating the correlation method. Passed to
\link[stats:cor]{stats::cor}. Also sets a name for the heatmap unless a \code{name} argument is
supplied.}

\item{...}{Other arguments passed to \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap}.}

\item{annotations}{A character vector indicating which column/row annotations
to include.}

\item{diagonal}{A character vector indicating what should happen to the
matrix diagonal. \code{"remove"} plots the diagonal in white, effectively
removing it from the plot. \code{"as_is"} plots the diagonal as is.}
}
\value{
A Heatmap plot of pairwise sample--sample correlations.
}
\description{
Compute and plot pairwise sample--sample correlations for a
\code{SummarizedExperiment} \code{assay}
}
\details{
If there are aesthetics set -- e.g. by \link{set_aesthetics} -- then
heatmap annotations are automatically created from those aesthetics.
Annotation position can be controlled by \code{annotations}.
}
