% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_genes.R
\name{plot_detected_genes}
\alias{plot_detected_genes}
\title{Plot detected gene counts for each library}
\usage{
plot_detected_genes(dds, threshold = 5)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{threshold}{Threshold for minimum normalized counts.}
}
\value{
A ggplot object bar plot of counts of detected genes for each library.
}
\description{
Generate a bar plot with the count of detected genes for each library.

If detected gene counts are not present in \code{dds} (i.e. no \code{detected_genes} column in \code{dds} column data),
detected genes are counted by \code{\link[=count_library_detected_genes]{count_library_detected_genes()}}.
}
