% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_size.R
\name{plot_library_sizes}
\alias{plot_library_sizes}
\title{Plot the library size for each sample}
\usage{
plot_library_sizes(se, library_size_var, interactive = TRUE)
}
\arguments{
\item{library_size_var}{A character of the variable in the
\code{SummarizedExperiment} sample data to plot. This variable must exist in the
\code{SummarizedExperiment} sample data.}

\item{interactive}{A logical controlling whether to return an interactive
plotly plot or a static ggplot.}
}
\value{
A dot plot of library size for each sample
}
\description{
Plot the library size for each sample
}
