% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_size.R
\name{plot_logcpm_density}
\alias{plot_logcpm_density}
\title{Plot the log CPM density of each sample}
\usage{
plot_logcpm_density(se, interactive = TRUE)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{interactive}{A logical controlling whether to return an interactive
plotly plot or a static ggplot.}
}
\value{
A density plot of the log CPM each sample
}
\description{
Plot the log CPM density of each sample
}
