% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-sample_data.R
\name{plot_modeling_variables_pairs}
\alias{plot_modeling_variables_pairs}
\title{Plot pairs scatter plot matrix for all modeling variables of
\code{SummarizedExperiment}}
\usage{
plot_modeling_variables_pairs(se, seed = 2025)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{seed}{A numeric seed to control the position of the point jittering.
Used for reproducible results.}
}
\value{
A scatter plot matrix comparing pairs of all sample variables set in
model variables.
}
\description{
Plot pairs scatter plot matrix for all modeling variables of
\code{SummarizedExperiment}
}
