% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary.R
\name{plot_results_summary}
\alias{plot_results_summary}
\title{Plot results summary data}
\usage{
plot_results_summary(res)
}
\arguments{
\item{res}{A \code{DESeqResults} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot results summary data
}
