% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_exclusions.R
\name{plot_sample_exclusions_heatmap}
\alias{plot_sample_exclusions_heatmap}
\title{Plot a heatmap of sample exclusions and the reasons for exclusions}
\usage{
plot_sample_exclusions_heatmap(se, sample_exclusions)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object}

\item{sample_exclusions}{A \code{list} of key:value pairs mapping reason for
exclusion to samples excluded.}
}
\value{
A \code{plotly} object heatmap of samples to be excluded and reasons for
exclusion
}
\description{
Plot a heatmap of sample exclusions and the reasons for exclusions
}
