% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-pca.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot a 2D scatter plot of PCA scores}
\usage{
plot_scores(
  pca,
  x = "PC1",
  y = "PC2",
  plot_type = c("plotly", "ggplot"),
  margin_plot = c("rug", "none"),
  label = "sample",
  color = NULL,
  shape = NULL
)
}
\arguments{
\item{pca}{A \code{PCA} object returned by \link{perform_pca}.}

\item{x}{A character vector of the principal component to plot on the x-axis.}

\item{y}{A character vector of the principal component to plot on the y-axis.}

\item{plot_type}{A character vector indicating the plot type to return.}

\item{margin_plot}{A character vector indicating the plot type to make in the
margins.}

\item{label}{A character vector indicating the sample data variable to use
for labeling samples. Labels don't show up on ggplots, but will be visible
on the plotly tooltip.}

\item{color}{A character vector indicating the sample data variable to use
for coloring the plot points.}

\item{shape}{A character vector indicating the sample data variable to use
for shape of the plot points.}
}
\value{
A plot of PC scores.
}
\description{
Plot a 2D scatter plot of PCA scores
}
