% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-pca.R
\name{plot_scree}
\alias{plot_scree}
\title{Plot a PCA scree plot}
\usage{
plot_scree(pca, plot_type = c("plotly", "ggplot"))
}
\arguments{
\item{pca}{A \code{PCA} object returned by \link{perform_pca}.}

\item{plot_type}{A character vector indicating the plot type to return.}
}
\value{
A scree plot.
}
\description{
Plot a PCA scree plot
}
