% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-counts.R
\name{read_counts_file}
\alias{read_counts_file}
\title{Read counts matrix from file}
\usage{
read_counts_file(path)
}
\arguments{
\item{path}{A string path to a counts file in CSV format such as the one
created by \href{https://github.com/utia-gc/rnaseq}{utia-gc/rnaseq}.}
}
\value{
An integer matrix with genes as rows and samples as columns. Row
names are gene names. Column names are sample names.
}
\description{
Read counts matrix from file
}
