% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_featureCounts_matrix.R
\name{read_featureCounts_matrix}
\alias{read_featureCounts_matrix}
\title{Read counts matrix from featureCounts file(s)}
\usage{
read_featureCounts_matrix(paths, pattern = "\\\\.bam")
}
\arguments{
\item{paths}{Paths to featureCounts files}

\item{pattern}{Pattern to look for and remove from sample names.
All instances of this pattern will be removed from sample names.

The default interpretation is a regular expression, as described in stringi::about_search_regex.}
}
\value{
A matrix of counts from featureCounts.
Columns correspond to samples with sample names as column names, and
rows correspond to features with feature names (content of \code{Geneid} column) as row names
}
\description{
Read counts matrix from featureCounts file(s)
}
