% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{run_if_changed}
\alias{run_if_changed}
\title{Wrapper around \code{\link[Hmisc:runifChanged]{Hmisc::runifChanged()}}.}
\usage{
run_if_changed(file, fun, ...)
}
\arguments{
\item{file}{Path to file to store cached results. \code{dirname(file)} is created
if it does not already exist.}

\item{fun}{The usually slow function to run.}

\item{...}{Additional arguments and objects passed to
\code{\link[Hmisc:runifChanged]{Hmisc::runifChanged()}}.}
}
\value{
Result of \code{fun}.
}
\description{
Wraps \code{\link[Hmisc:runifChanged]{Hmisc::runifChanged()}}. Introduces convenience whereby
the directory for \code{file} is created if it doesn't already exist.
}
