% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{set_aesthetics}
\alias{set_aesthetics}
\title{Set aesthetics metadata in a \code{SummarizedExperiment}}
\usage{
set_aesthetics(se, variable, values, aesthetic = c("color", "shape"))
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{variable}{A character vector for the sample data variable to set
aesthetic values for.}

\item{values}{A named vector mapping values in the sample data \code{variable} to
values of an aesthetic.}

\item{aesthetic}{A vector indicating the aesthetic to set.}
}
\value{
The \code{SummarizedExperiment} with the aesthetics set in the metadata
slot.
}
\description{
Set aesthetics metadata in a \code{SummarizedExperiment}
}
