% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{set_formula}
\alias{set_formula}
\title{Set a formula in the metadata of a \code{SummarizedExperiment}}
\usage{
set_formula(se, name, formula)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{name}{A character name for the formula.}

\item{formula}{A formula}
}
\value{
The \code{SummarizedExperiment} with the formula added to the metadata
slot.
}
\description{
Set a formula in the metadata of a \code{SummarizedExperiment}
}
