% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-se.R
\name{set_modeling_variable}
\alias{set_modeling_variable}
\title{Set a modeling variable in the metadata of a \code{SummarizedExperiment}}
\usage{
set_modeling_variable(
  se,
  variable = c("interest", "nuisance", "screening"),
  values
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{variable}{A character vector for which variable to set}

\item{values}{A character vector of which variables to set the modeling
variable as}
}
\value{
The \code{SummarizedExperiment} with the modeling variable set to values
in the metadata slot
}
\description{
Set a modeling variable in the metadata of a \code{SummarizedExperiment}
}
\details{
The character vectors in \code{values} must all be names of variables
available in the sample data. Otherwise, an \code{error_no_sample_variable} is
thrown.
}
