% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-sample_data.R
\name{tbl_interest_variables_freq}
\alias{tbl_interest_variables_freq}
\title{Display a frequency table of number of samples per combination of interest
variables in a \code{SummarizedExperiment}}
\usage{
tbl_interest_variables_freq(se, table_format = c("datatable", "data.frame"))
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{table_format}{A character vector of the frequency table format to
return}
}
\value{
A frequency table of number of samples per combination of interest
variables
}
\description{
Display a frequency table of number of samples per combination of interest
variables in a \code{SummarizedExperiment}
}
