% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_size.R
\name{tbl_library_sizes}
\alias{tbl_library_sizes}
\title{Make table of library sizes from a matrix of counts}
\usage{
tbl_library_sizes(counts, interactive = TRUE)
}
\arguments{
\item{counts}{An integer matrix with genes as rows and samples as columns.
Row names are gene names. Column names are sample names.}

\item{interactive}{A logical controlling whether to return an interactive
data table with \code{\link[DT:datatable]{DT::datatable()}} or a \code{data.frame} from \link[tibble:tbl_df-class]{tibble::tbl_df}.}
}
\value{
A table of library sizes (matrix column sums) for each sample ordered
by increasing library size. Returns an interactive data table with
\code{\link[DT:datatable]{DT::datatable()}} or a \code{data.frame} from \link[tibble:tbl_df-class]{tibble::tbl_df}.
}
\description{
Make table of library sizes from a matrix of counts
}
