% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-sample_data.R
\name{tbl_sample_data}
\alias{tbl_sample_data}
\title{Display table of sample data}
\usage{
tbl_sample_data(se, interactive = TRUE)
}
\arguments{
\item{interactive}{A logical controlling whether to return an interactive
data table with \code{\link[DT:datatable]{DT::datatable()}} or a \code{data.frame} from \link[tibble:tbl_df-class]{tibble::tbl_df}.}
}
\value{
A table of sample data. Returns an interactive data table with
\code{\link[DT:datatable]{DT::datatable()}} or a \code{data.frame} from \link[tibble:tbl_df-class]{tibble::tbl_df}.
}
\description{
Display table of sample data
}
