% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{TestSpecs}
\alias{TestSpecs}
\title{Specify tests for differential analysis}
\usage{
TestSpecs(path, alpha, lfc_threshold)
}
\arguments{
\item{path}{A character path to YAML of test specs}

\item{alpha}{A numeric alpha threshold for differential features}

\item{lfc_threshold}{A numeric log2 fold change threshold for differential features}
}
\value{
A \code{TestSpecs} object
}
\description{
Specify tests for differential analysis
}
