% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_specs.R
\name{validate_TestSpecs}
\alias{validate_TestSpecs}
\title{Validate a \code{TestSpecs} object}
\usage{
validate_TestSpecs(test_specs)
}
\arguments{
\item{test_specs}{A \code{TestSpecs} object}
}
\value{
(Invisibly) a \code{TestSpecs} object
}
\description{
Validate a \code{TestSpecs} object
}
