% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-se.R
\name{warn_no_overwrite}
\alias{warn_no_overwrite}
\title{Warn that a file was not overwritten}
\usage{
warn_no_overwrite(path)
}
\arguments{
\item{path}{A character path to an existing file or directory}
}
\value{
(Invisible) \code{NULL}
}
\description{
Warn that a file was not overwritten
}
