% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-se.R
\name{write_se}
\alias{write_se}
\title{Write a \code{SummarizedExperiment} object to file as an RDS.}
\usage{
write_se(se, path, overwrite = FALSE)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{path}{A character path to write the \code{SummarizedExperiment} file to}

\item{overwrite}{A logical controlling whether a file should be overwritten
if \code{path} already exists}
}
\value{
(Invisible) a logical indicating whether the \code{se} was successfully
written to \code{path}
}
\description{
Write a \code{SummarizedExperiment} object to file as an RDS.
}
\examples{
# don't run
if (FALSE) {
  # create SummarizedExperiment object
  counts <- random_matrix()
  col_data <- random_col_data()
  se <- SummarizedExperiment::SummarizedExperiment(
    assays = list(counts = counts),
    colData = col_data
  )

  # write the SummarizedExperiment
  write_se(se, path = "se.rds")
}
}
