% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats-pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform principal components analysis (PCA) on an assay from
\code{SummarizedExperiment}}
\usage{
perform_pca(se, assay, ntop = 500)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}}

\item{assay}{A character vector for the assay from \code{SummarizedExperiment} to
use as input to the PCA.}

\item{ntop}{An integer for the number of most variable genes to use as input
to the PCA.}
}
\value{
A list of type \code{PCA}. Contains PC scores joined to sample data, a
data.frame of proportion of variance explained by each PC, the \code{prcomp}
object returned by \link[stats:prcomp]{stats::prcomp}, and the metadata slot from the
\code{SummarizedExperiment}.
}
\description{
Perform PCA for an assay from \code{SummarizedExperiment}, and return
the results alongside sample data from the \code{SummarizedExperiment}.
}
\details{
\code{perform_pca()} is essentially a convenience around running
\link[stats:prcomp]{stats::prcomp} on a particular assay matrix from a \code{SummarizedExperiment},
joining the PC scores to the sample data, and also returning some data
about the proportion variance explained.

Note that there is no additional transformation of the expression data
aside from subsetting the expression matrix to include only the \code{ntop} most
variable genes and coercing the matrix to the expected shape for
\link[stats:prcomp]{stats::prcomp}. If you wish to center or scale data, for example, you must
do so prior to running \code{perform_pca()} and store the transformed matrix in
a new element in the \code{assay} slot of your \code{SummarizedExperiment}.
}
