% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_scree}
\alias{plot_scree}
\title{Make a scree plot for a PCA DFrame}
\usage{
plot_scree(pca_df)
}
\arguments{
\item{pca_df}{A \code{Dframe} of PCA data}
}
\value{
A ggplot object scree plot
}
\description{
Make a scree plot for a PCA DFrame
}
